+++ api/pom.xml
@@ -18,6 +18,7 @@
 		<spring-security.version>6.1.2</spring-security.version>
 	</properties>
 	<dependencies>
+
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter</artifactId>
