+++ api/src/main/java/com/csiro/tickets/models/Label.java
@@ -1,4 +1,4 @@
-package com.csiro.snomio.models.tickets;
+package com.csiro.tickets.models;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -6,12 +6,10 @@ import jakarta.persistence.EntityListeners;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import jakarta.persistence.Version;
 import java.time.Instant;
 import lombok.Data;
-import org.hibernate.envers.Audited;
 import org.springframework.data.annotation.CreatedBy;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedBy;
@@ -20,10 +18,9 @@ import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
 @Entity
 @Data
-@Table(name="attachment")
-@Audited
+@Table(name="label")
 @EntityListeners(AuditingEntityListener.class)
-public class Attachment {
+public class Label {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -31,9 +28,6 @@ public class Attachment {
 
   @Version
   private Integer version;
-
-  @ManyToOne
-  private Ticket ticket;
 
   @Column(name = "created", nullable = false, updatable = false)
   @CreatedDate
@@ -50,21 +44,6 @@ public class Attachment {
   @Column(name = "modified_by")
   @LastModifiedBy
   private String modifiedBy;
-
-  @Column
-  private String description;
-
-  @Column
-  private String data;
-
-  @Column
-  private Integer length;
-
-  @Column
-  private String sha256;
-
-  @ManyToOne
-  private AttachmentType attachmentType;
 
 
 }
