+++ api/src/main/java/com/csiro/tickets/models/TicketType.java
@@ -1,30 +1,28 @@
-package com.csiro.snomio.models.tickets;
+package com.csiro.tickets.models;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
+import jakarta.persistence.EntityListeners;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.Table;
 import jakarta.persistence.Version;
 import java.time.Instant;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Data;
-import lombok.NoArgsConstructor;
 import org.hibernate.envers.Audited;
 import org.springframework.data.annotation.CreatedBy;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedBy;
 import org.springframework.data.annotation.LastModifiedDate;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
 @Entity
 @Data
-@NoArgsConstructor
-@AllArgsConstructor
+@Table(name="attachment")
 @Audited
-@Table(name="transition")
-public class Transition {
+@EntityListeners(AuditingEntityListener.class)
+public class TicketType {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -51,4 +49,8 @@ public class Transition {
 
   @Column
   private String name;
+
+  @Column
+  private String description;
+
 }
