+++ api/src/main/java/com/csiro/tickets/models/Transition.java
@@ -1,19 +1,14 @@
-package com.csiro.snomio.models.tickets;
+package com.csiro.tickets.models;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.OneToMany;
-import jakarta.persistence.OrderBy;
 import jakarta.persistence.Table;
 import jakarta.persistence.Version;
 import java.time.Instant;
-import java.util.List;
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.hibernate.envers.Audited;
@@ -23,13 +18,12 @@ import org.springframework.data.annotation.LastModifiedBy;
 import org.springframework.data.annotation.LastModifiedDate;
 
 @Entity
-@Builder
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 @Audited
-@Table(name="ticket")
-public class Ticket {
+@Table(name="transition")
+public class Transition {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -55,15 +49,5 @@ public class Ticket {
   private String modifiedBy;
 
   @Column
-  private String title;
-
-  @Column
-  private String description;
-
-  @ManyToOne
-  private TicketType ticketType;
-
-  @ManyToOne
-  private State state;
+  private String name;
 }
-
